CSP-2540 Refactored controller logic and add query result tests
